SLSTR Level1B processor code
The code works. Needs an API authentication. Fundamentally, it
downloads the products, unzips the files, and grids the BT to a
sensible equal Area Projection. There are likely issues with
missing values, but this should be reasonably complete. There
is also not a lot of diagnostics or robustness if the EUMETSAT
server decides to barf...